+++ R/tm_g_forest_rsp.R
@@ -4,6 +4,8 @@
 #'
 #' @param subgroup_var \code{\link[teal]{choices_selected}} object with all available choices and preselected option
 #' for variable names that can be used as the default subgroups
+#' @param fixed_symbol_size (\code{logical}) When (\code{TRUE}), the same symbol size is used for plotting each
+#' estimate. Otherwise, the symbol size will be proportional to the sample size in each each subgroup.
 #' @param plot_height vector with three elements defining selected, min and max plot height
 #' @param cex multiplier applied to overall font size
 #'
@@ -17,22 +19,25 @@
 #' library(dplyr)
 #'
 #' ADSL <- radsl(cached = TRUE)
-#' ADRS <- radrs(ADSL, cached = TRUE) %>% dplyr::filter(AVISIT == "FOLLOW UP")
+#' ADRS <- radrs(cached = TRUE) %>%
+#'      dplyr::filter(PARAMCD %in% c("BESRSPI", "INVET"))
 #'
 #' app <- init(
 #'   data = cdisc_data(
 #'    cdisc_dataset("ADSL", ADSL),
 #'    cdisc_dataset("ADRS", ADRS),
 #'    code = 'ADSL <- radsl(cached = TRUE)
-#'            ADRS <- radrs(ADSL, cached = TRUE) %>% dplyr::filter(AVISIT == "FOLLOW UP")',
+#'            ADRS <- radrs(cached = TRUE) %>%
+#'            dplyr::filter(PARAMCD %in% c("BESRSPI", "INVET"))',
 #'    check = FALSE),
 #'   modules = root_modules(
 #'     tm_g_forest_rsp(
 #'       label = "Forest Response",
 #'       dataname = "ADRS",
 #'       arm_var = choices_selected(c("ARM", "ARMCD"), "ARM"),
-#'       paramcd = choices_selected(levels(ADRS$PARAMCD), "OVRINV"),
-#'       subgroup_var = choices_selected(names(ADSL), c("RACE", "SEX")),
+#'       paramcd = choices_selected(c("BESRSPI", "INVET"), "BESRSPI"),
+#'       subgroup_var = choices_selected(names(ADSL), c("BMRKR2", "SEX")),
+#'       strata_var = choices_selected(c("STRATA1", "STRATA2"), "STRATA2"),
 #'       plot_height = c(600L, 200L, 2000L)
 #'     )
 #'   )
@@ -46,6 +51,9 @@ tm_g_forest_rsp <- function(label,
                             arm_var,
                             paramcd,
                             subgroup_var,
+                            strata_var,
+                            conf_int = choices_selected(c(0.8, 0.85, 0.90, 0.95, 0.99, 0.995), 0.95, keep_order = TRUE),
+                            fixed_symbol_size = TRUE,
                             plot_height = c(700L, 200L, 2000L),
                             cex = 1.3,
                             pre_output = helpText("graph needs to be of a certain width to be displayed"),
@@ -56,6 +64,9 @@ tm_g_forest_rsp <- function(label,
   stopifnot(is.choices_selected(arm_var))
   stopifnot(is.choices_selected(paramcd))
   stopifnot(is.choices_selected(subgroup_var))
+  stopifnot(is.choices_selected(strata_var))
+  stopifnot(is.choices_selected(conf_int))
+  stopifnot(is_logical_single(fixed_symbol_size))
   stop_if_not(list(
     is_integer_vector(plot_height) && length(plot_height) == 3,
     "plot_height should be vector of three integers specyfing selected height, min and max height"
@@ -103,10 +114,12 @@ ui_g_forest_rsp <- function(id, ...) {
       helpText("Analysis data:", tags$code(a$dataname)),
       optionalSelectInput(
         ns("paramcd"),
-        div("PARAMCD", tags$br(), helpText("Select one type of response to analyze.")),
+        "PARAMCD",
         a$paramcd$choices,
         a$paramcd$selected,
-        multiple = FALSE
+        multiple = FALSE,
+        fixed = a$paramcd$fixed,
+        label_help = helpText("Select one type of response to analyze.")
       ),
       selectInput(
         ns("responders"),
@@ -120,7 +133,8 @@ ui_g_forest_rsp <- function(id, ...) {
         "Arm Variable",
         a$arm_var$choices,
         a$arm_var$selected,
-        multiple = FALSE
+        multiple = FALSE,
+        fixed = a$arm_var$fixed
       ),
       selectInput(
         ns("ref_arm"),
@@ -144,7 +158,16 @@ ui_g_forest_rsp <- function(id, ...) {
         a$subgroup_var$choices,
         a$subgroup_var$selected,
         multiple = TRUE,
-        label_help = helpText("are taken from", tags$code("ADSL"))
+        label_help = helpText("are taken from", tags$code("ADSL")),
+        fixed = a$subgroup_var$fixed
+      ),
+      optionalSelectInput(ns("strata_var"),
+                          "Stratify by",
+                          a$strata_var$choices,
+                          a$strata_var$selected,
+                          multiple = TRUE,
+                          label_help = helpText("from ", tags$code("ADSL")),
+                          fixed = a$strata_var$fixed
       ),
       tags$label(
         "Plot Settings",
@@ -162,7 +185,21 @@ ui_g_forest_rsp <- function(id, ...) {
         "plot width",
         c(980L, 500L, 2000L),
         ticks = FALSE
+      ),
+      panel_group(
+        panel_item(
+        "Additional plot settings",
+        optionalSelectInput(
+          ns("conf_int"),
+          "Level of Confidence",
+          a$conf_int$choices,
+          a$conf_int$selected,
+          multiple = FALSE,
+          fixed = a$conf_int$fixed
+        ),
+        checkboxInput(ns("fixed_symbol_size"), "Fixed symbol size", value = TRUE)
       )
+     )
     ),
     forms = actionButton(
       ns("show_rcode"),
@@ -222,11 +259,18 @@ srv_g_forest_rsp <- function(input,
     ref_arm <- input$ref_arm
     comp_arm <- input$comp_arm
     subgroup_var <- input$subgroup_var
+    strata_var <- input$strata_var
+    conf_int <- as.numeric(input$conf_int)
+    col_symbol_size <- if (input$fixed_symbol_size) {
+      NULL
+    } else {
+      1
+    }
 
     # validate your input values
     validate_standard_inputs(
       adsl = adsl_filtered,
-      adslvars = c("USUBJID", "STUDYID", arm_var, subgroup_var),
+      adslvars = c("USUBJID", "STUDYID", arm_var, subgroup_var, strata_var),
       anl = anl_filtered,
       anlvars = c("USUBJID", "STUDYID",  "PARAMCD", "AVAL", "AVALC"),
       arm_var = arm_var,
@@ -236,6 +280,11 @@ srv_g_forest_rsp <- function(input,
 
     validate_in(responders, anl_filtered$AVALC, "Responder values cannot be found in AVALC.")
     validate_in(paramcd, anl_filtered$PARAMCD, "Response parameter cannot be found in PARAMCD.")
+    validate(
+      need(length(conf_int) == 1, "Please select level of confidence."),
+      need(all(vapply(adsl_filtered[, subgroup_var], is.factor, logical(1))),
+           "Not all subgroup variables are factors.")
+    )
 
     # perform analysis
     anl_name <- paste0(dataname, "_FILTERED")
@@ -243,15 +292,17 @@ srv_g_forest_rsp <- function(input,
     adsl_name <- "ADSL_FILTERED"
     assign(adsl_name, adsl_filtered)
 
-    adsl_vars <- unique(c("USUBJID", "STUDYID", arm_var, subgroup_var))
-    anl_vars <- c("USUBJID", "STUDYID", "AVALC")
+    adsl_vars <- unique(c("USUBJID", "STUDYID", arm_var, subgroup_var, strata_var)) # nolint
+    anl_vars <- c("USUBJID", "STUDYID", "AVALC") # nolint
 
     chunks_reset(envir = environment())
 
     chunks_push(bquote({
+
       adsl_p <- subset(.(as.name(adsl_name)), .(as.name(arm_var)) %in% c(.(ref_arm), .(comp_arm)))
       anl_p <- subset(.(as.name(anl_name)), PARAMCD %in% .(paramcd))
 
+      adsl_p[, .(subgroup_var)] <- droplevels(adsl_p[, .(subgroup_var)])
       anl <- merge(adsl_p[, .(adsl_vars)], anl_p[, .(anl_vars)],
                    all.x = FALSE, all.y = FALSE, by = c("USUBJID", "STUDYID"))
 
@@ -285,14 +336,16 @@ srv_g_forest_rsp <- function(input,
         } else {
           bquote(NULL)
         }),
+        strata_data = if (!is.null(.(strata_var))) {
+          anl[, .(strata_var), drop = FALSE]
+        } else {
+          NULL
+        },
+        conf_int = .(conf_int),
         total = "All Patients",
         dense_header = TRUE
       )
-    }))
 
-    chunks_safe_eval()
-
-    chunks_push(quote({
       row.names(tbl) <- sapply(
         row.names(tbl),
         function(x) {
@@ -301,21 +354,32 @@ srv_g_forest_rsp <- function(input,
       )
     }))
 
-    chunks_push(call(
-      "g_forest",
-      tbl = quote(tbl),
-      col_x = 8,
-      col_ci = 9,
-      vline = 1,
-      forest_header = bquote(paste0(levels(anl[[.(arm_var)]]), "\nbetter")),
-      xlim = c(.1, 10),
-      logx = TRUE,
-      x_at = c(.1, 1, 10)
-    ))
+    chunks_safe_eval()
 
-    p <- chunks_safe_eval()
+    chunks_push(bquote({
+      p <- g_forest(
+        tbl = tbl,
+        col_x = 8,
+        col_ci = 9,
+        vline = 1,
+        forest_header = paste0(levels(anl[[.(arm_var)]]), "\nbetter"),
+        xlim = c(.1, 10),
+        logx = TRUE,
+        x_at = c(.1, 1, 10),
+        draw = FALSE,
+        col_symbol_size = .(col_symbol_size)
+      )
+      if (!is.null(footnotes(p))) {
+        p <- decorate_grob(p, title = "Forest plot", footnotes = footnotes(p),
+                           gp_footnotes = gpar(fontsize = 12))
+      }
+      grid.newpage()
+      grid.draw(p)
+    }))
 
-    p
+
+    chunks_safe_eval()
+
   })
 
   ## dynamic plot height
@@ -323,7 +387,8 @@ srv_g_forest_rsp <- function(input,
     plot_height <- input$plot_height
     plot_width <- input$plot_width
     validate(need(plot_height, "need valid plot height"))
-    div(style = 'overflow-x: scroll',
+    validate(need(plot_width, "need valid plot width"))
+    div(style = "overflow-x: scroll",
         plotOutput(session$ns("forest_plot"),
                    height = plot_height,
                    width = plot_width)