+++ R/tm_g_km.R
@@ -36,7 +36,7 @@
 #' app <- init(
 #'   data = cdisc_data(
 #'     cdisc_dataset("ADSL", ADSL), cdisc_dataset("ADTTE", ADTTE),
-#'     code = "radsl(cached = TRUE)
+#'     code = "ADSL <- radsl(cached = TRUE)
 #'             ADTTE <- radtte(ADSL, cached = TRUE)",
 #'     check = FALSE
 #'   ),
@@ -62,6 +62,7 @@ tm_g_km <- function(label,
                     paramcd,
                     facet_var,
                     strata_var,
+                    conf_int = choices_selected(c(0.8, 0.85, 0.90, 0.95, 0.99, 0.995), 0.95),
                     plot_height = c(1200, 400, 5000),
                     pre_output = helpText("x-axes for different factes may not have the same scale"),
                     post_output = NULL) {
@@ -101,14 +102,16 @@ ui_g_km <- function(id, ...) {
         "Arm Variable",
         choices = a$arm_var$choices,
         selected = a$arm_var$selected,
-        multiple = FALSE
+        multiple = FALSE,
+        fixed = a$arm_var$fixed
       ),
       optionalSelectInput(
         ns("paramcd"),
         "Time to Event (Endpoint)",
         choices = a$paramcd$choices,
         selected = a$paramcd$selected,
-        multiple = FALSE
+        multiple = FALSE,
+        fixed = a$paramcd$fixed
       ),
       optionalSelectInput(
         ns("strata_var"),
@@ -116,7 +119,8 @@ ui_g_km <- function(id, ...) {
         choices = a$strata_var$choices,
         selected = a$strata_var$selected,
         multiple = TRUE,
-        label_help = helpText("currently taken from ADSL")
+        label_help = helpText("currently taken from ADSL"),
+        fixed = a$strata_var$fixed
       ),
       optionalSelectInput(
         ns("facet_var"),
@@ -124,7 +128,8 @@ ui_g_km <- function(id, ...) {
         choices = a$facet_var$choices,
         selected = a$facet_var$selected,
         multiple = TRUE,
-        label_help = helpText("currently taken from ADSL")
+        label_help = helpText("currently taken from ADSL"),
+        fixed = a$facet_var$fixed
       ),
       selectInput(
         ns("ref_arm"),
@@ -153,7 +158,7 @@ ui_g_km <- function(id, ...) {
         panel_item(
           "Additional plot settings",
           numericInput(
-            inputId = ns('font_size'),
+            inputId = ns("font_size"),
             label = "Font size",
             value = 8,
             min = 5,
@@ -162,13 +167,13 @@ ui_g_km <- function(id, ...) {
             width = "100%"
           ),
           checkboxInput(
-            inputId = ns('show_km_table'),
+            inputId = ns("show_km_table"),
             label = "Show KM table",
             value = TRUE,
             width = "100%"
           ),
           checkboxInput(
-            inputId = ns('show_coxph_table'),
+            inputId = ns("show_coxph_table"),
             label = "Show CoxPH table",
             value = TRUE,
             width = "100%"
@@ -176,8 +181,15 @@ ui_g_km <- function(id, ...) {
           radioButtons(
             ns("pval_method"),
             "p-value method",
-            choices = c("wald", "logrank", "likelihood"),
-            selected = "wald"
+            choices = c("wald", "log-rank", "likelihood"),
+            selected = "log-rank"
+          ),
+          optionalSelectInput(ns("conf_int"),
+                              "Level of Confidence",
+                              a$conf_int$choices,
+                              a$conf_int$selected,
+                              multiple = FALSE,
+                              fixed = a$conf_int$fixed
           ),
           textInput(ns("xlab"), "X-axis label", "Overall survival in ")
         )
@@ -221,18 +233,19 @@ srv_g_km <- function(input,
     anl_filtered <- datasets$get_data(dataname, filtered = TRUE, reactive = TRUE)
     ADSL_FILTERED <- datasets$get_data("ADSL", reactive = TRUE, filtered = TRUE) # nolint
 
-    paramcd <- input$paramcd
+    paramcd <- input$paramcd # nolint
     arm_var <- input$arm_var
     facet_var <- input$facet_var
     ref_arm <- input$ref_arm
     comp_arm <- input$comp_arm
     strata_var <- input$strata_var
     combine_comp_arms <- input$combine_comp_arms
-    pval_method <- input$pval_method
-    xlab <- input$xlab
-    tbl_fontsize <- input$font_size
-    if_show_km <- input$show_km_table
-    if_show_coxph <- input$show_coxph_table
+    pval_method <- input$pval_method # nolint
+    conf_int <- as.numeric(input$conf_int) # nolint
+    xlab <- input$xlab # nolint
+    tbl_fontsize <- input$font_size # nolint
+    if_show_km <- input$show_km_table # nolint
+    if_show_coxph <- input$show_coxph_table # nolint
 
     if (length(facet_var) == 0) {
       facet_var <<- NULL
@@ -296,7 +309,7 @@ srv_g_km <- function(input,
     chunks_push(bquote({
       formula_coxph <- as.formula(
         .(paste0(
-          "Surv(AVAL, 1-CNSR) ~ ", arm_var,
+          "Surv(AVAL, 1-CNSR) ~ arm(", arm_var, ")",
           ifelse(is.null(strata_var), "", paste0(" + strata(", paste(strata_var, collapse = ","), ")"))
         ))
       )
@@ -313,14 +326,14 @@ srv_g_km <- function(input,
 
     if (is.null(facet_var)) {
       chunks_push(bquote({
-        fit_km <- survfit(formula_km, data = anl, conf.type = "plain")
+        fit_km <- survfit(formula_km, data = anl, conf.int = .(conf_int), conf.type = "plain")
         grid.newpage()
         p <- g_km(fit_km = fit_km, col = NA, draw = FALSE, xlab = paste(.(xlab), time_unit))
 
         .(
           if (if_show_km) {
             bquote({
-              tbl_km <- t_km(fit_km)
+              tbl_km <- t_km(formula_km, data = anl, conf.int = .(conf_int), conf.type = "plain")
               km_grob <- textGrob(
                 label = toString(tbl_km, gap = 1),
                 x = unit(1, "npc") - stringWidth(toString(tbl_km, gap = 1)) - unit(1, "lines"),
@@ -337,8 +350,13 @@ srv_g_km <- function(input,
         .(
           if (if_show_coxph) {
             bquote({
-              fit_coxph <- coxph(formula_coxph, data = anl, ties = "exact")
-              tbl_coxph <- t_coxph(fit_coxph, pval_method = pval_method)
+              tbl_coxph <- t_coxph(
+                formula_coxph,
+                data = anl,
+                conf.int = .(conf_int),
+                pval_method = .(pval_method),
+                ties = "exact"
+              )
               text_coxph <- paste0(info_coxph, "\n", toString(tbl_coxph, gap = 1))
               coxph_grob <- textGrob(
                 label = text_coxph, x = unit(1, "lines"), y = unit(1, "lines"),
@@ -375,7 +393,7 @@ srv_g_km <- function(input,
             textGrob(paste0("Less than 5 patients in ", label, " group"))
           } else {
             x[[.(arm_var)]] <- factor(x[[.(arm_var)]])
-            fit_km <- survfit(formula_km, data = x, conf.type = "plain")
+            fit_km <- survfit(formula_km, data = x, conf.int = .(conf_int), conf.type = "plain")
             p <- g_km(
               fit_km = fit_km, col = NA, title = paste0("Kaplan - Meier Plot for: ", label),
               xticks = xticks, draw = FALSE, xlab = paste(.(xlab), time_unit)
@@ -384,7 +402,7 @@ srv_g_km <- function(input,
             .(
               if (if_show_km) {
                 bquote({
-                  tbl_km <- t_km(fit_km)
+                  tbl_km <- t_km(formula_km, data = x, conf.int = .(conf_int), conf.type = "plain")
                   km_grob <- textGrob(
                     label = toString(tbl_km, gap = 1),
                     x = unit(1, "npc") - stringWidth(toString(tbl_km, gap = 1)) - unit(1, "lines"),
@@ -401,8 +419,13 @@ srv_g_km <- function(input,
             .(
               if (if_show_coxph) {
                 bquote({
-                  fit_coxph <- coxph(formula_coxph, data = x, ties = "exact")
-                  tbl_coxph <- t_coxph(fit_coxph, pval_method = .(pval_method))
+                  tbl_coxph <- t_coxph(
+                    formula_coxph,
+                    data = x,
+                    conf.int = .(conf_int),
+                    pval_method = .(pval_method),
+                    ties = "exact"
+                  )
                   text_coxph <- paste0(info_coxph, "\n", toString(tbl_coxph, gap = 1))
                   coxph_grob <- textGrob(
                     label = text_coxph,