+++ R/tm_t_rsp.R
@@ -50,7 +50,7 @@
 #'       dataname = 'ADRS',
 #'       arm_var = choices_selected(c("ARM", "ARMCD"), "ARM"),
 #'       paramcd = choices_selected(levels(ADRS$PARAMCD), "BESRSPI"),
-#'       strata_var = choices_selected(c("SEX", "BMRKR2"), "SEX")
+#'       strata_var = choices_selected(c("SEX", "BMRKR2"), NULL)
 #'     )
 #'   )
 #' )
@@ -121,10 +121,12 @@ ui_t_rsp <- function(id, ...) {
       #Response related parameters
       optionalSelectInput(
         ns("paramcd"),
-        div("PARAMCD", tags$br(), helpText("Select one type of response to analyze.")),
+        "PARAMCD",
         choices = a$paramcd$choices,
         selected = a$paramcd$selected,
-        multiple = FALSE
+        multiple = FALSE,
+        fixed = a$paramcd$fixed,
+        label_help = helpText("Select one type of response to analyze.")
       ),
       selectInput(
         ns("responders"),
@@ -139,7 +141,8 @@ ui_t_rsp <- function(id, ...) {
         "Arm Variable",
         choices = a$arm_var$choices,
         selected = a$arm_var$selected,
-        multiple = FALSE
+        multiple = FALSE,
+        fixed = a$arm_var$fixed
       ),
       selectInput(
         ns("ref_arm"),
@@ -168,7 +171,8 @@ ui_t_rsp <- function(id, ...) {
         choices = a$strata_var$choices,
         selected = a$strata_var$selected,
         multiple = TRUE,
-        label_help = helpText("taken from:", tags$code("ADSL"))
+        label_help = helpText("taken from:", tags$code("ADSL")),
+        fixed = a$strata_var$fixed
       )
     ),
     forms = actionButton(
@@ -276,8 +280,8 @@ srv_t_rsp <- function(input,
     adsl_name <- "ADSL_FILTERED"
     assign(adsl_name, adsl_filtered)
 
-    adsl_vars <- unique(c("USUBJID", "STUDYID", arm_var, strata_var))
-    anl_vars <- c("USUBJID", "STUDYID", "AVAL", "AVALC", "PARAMCD")
+    adsl_vars <- unique(c("USUBJID", "STUDYID", arm_var, strata_var)) # nolint
+    anl_vars <- c("USUBJID", "STUDYID", "AVAL", "AVALC", "PARAMCD") # nolint
 
 
     chunks_reset(envir = environment())