+++ R/tm_t_summary.R
@@ -73,13 +73,17 @@ ui_t_summary <- function(id, ...) {
                           "Arm Variable",
                           a$arm_var$choices,
                           a$arm_var$selected,
-                          multiple = FALSE),
+                          multiple = FALSE,
+                          fixed = a$arm_var$fixed
+      ),
       checkboxInput(ns("add_total"), "Add All Patients column", value = TRUE),
       optionalSelectInput(ns("summarize_vars"),
                           "Summarize Variables",
                           a$summarize_vars$choices,
                           a$summarize_vars$selected,
-                          multiple = TRUE)
+                          multiple = TRUE,
+                          fixed = a$summarize_vars$fixed
+      )
     ),
     forms = actionButton(ns("show_rcode"), "Show R Code", width = "100%"),
     pre_output = a$pre_output,
@@ -129,7 +133,7 @@ srv_t_summary <- function(input, output, session, datasets, dataname) {
       )
     )
 
-    total <- if (add_total) "All Patients" else NULL
+    total <- if (add_total) "All Patients" else NULL # nolint
 
     chunks_push(bquote({
       tbl <- t_summary(