+++ R/tm_t_tte.R
@@ -20,6 +20,8 @@
 #' variable names that can be used as \code{PARAMCD} variable
 #' @param strata_var \code{\link[teal]{choices_selected}} object with all available choices and preselected option
 #' for variable names that can be used for stratification
+#' @param conf_int \code{\link[teal]{choices_selected}} object with all available choices and preselected option
+#' for variable names that can be used for confidence level for computation of the confidence intervals.
 #' @param time_points \code{\link[teal]{choices_selected}} object with all available choices and preselected option
 #' for variable names that can be used \code{\link[tern]{t_tte}}
 #' @param time_unit (\code{character}) with unit of \code{dataname$AVAL}, please use singular e.g. month instead
@@ -53,6 +55,16 @@
 #' ADSL <- radsl(cached = TRUE)
 #' ADTTE <- radtte(ADSL, cached = TRUE)
 #'
+#' arm_ref_comp = list(
+#'   ACTARMCD = list(
+#'     ref = "ARM B",
+#'     comp = c("ARM A", "ARM C")
+#'   ),
+#'    ARM = list(
+#'     ref = "B: Placebo",
+#'     comp = c("A: Drug X", "C: Combination")
+#'   )
+#' )
 #'
 #' app <- init(
 #'     data = cdisc_data(cdisc_dataset("ADSL", ADSL), cdisc_dataset("ADTTE", ADTTE),
@@ -63,61 +75,13 @@
 #'         tm_t_tte(
 #'             label = "Time To Event Table",
 #'             dataname = 'ADTTE',
-#'             arm_var = choices_selected(c("ARM", "ARMCD"), "ARM"),
-#'             paramcd = choices_selected(unique(ADTTE$PARAMCD), "OS"),
+#'             arm_var = choices_selected(c("ARM", "ARMCD", "ACTARMCD"), "ARM"),
+#'             arm_ref_comp = arm_ref_comp,
+#'             paramcd = choices_selected(value_choices(ADTTE, "PARAMCD", "PARAM"), "OS"),
 #'             strata_var = choices_selected(c("SEX", "BMRKR2"), "SEX"),
 #'             time_points = choices_selected(c(6, 8), 6),
 #'             time_unit = "month",
 #'             event_desc_var = "EVNTDESC"
-#'         )
-#'     )
-#' )
-#'
-#'
-#' \dontrun{
-#'  shinyApp(app$ui, app$server)
-#' }
-#'
-#' ## Define default reference & comparison arms based on
-#' ## ARM variable
-#' library(teal)
-#' library(dplyr)
-#' library(magrittr)
-#'
-#' ADSL <- mutate(radsl(seed = 1),
-#'   ARM1 = sample(c("DUMMY A", "DUMMY B"),
-#'   n(), TRUE))
-#' attr(ADSL$ARM1, "label") <- "dummy"
-#' ADTTE <- radtte(ADSL, seed = 1)
-#'
-#'
-#' arm_ref_comp = list(
-#'   ACTARMCD = list(
-#'     ref = "ARM A",
-#'     comp = c("ARM B", "ARM C")
-#'   ),
-#'   ARM1 = list(
-#'     ref = "DUMMY B",
-#'     comp = "DUMMY A"
-#'   )
-#' )
-#' app <- init(
-#'     data = cdisc_data(cdisc_dataset("ADSL", ADSL), cdisc_dataset("ADTTE", ADTTE),
-#'         code = "ADSL <- radsl(seed = 1) %>%
-#'                 mutate(., ARM1 = sample(c('DUMMY A', 'DUMMY B'), n(), TRUE))
-#'                 ADTTE <- radtte(ADSL, seed = 1)",
-#'         check = FALSE),
-#'     modules = root_modules(
-#'         tm_t_tte(
-#'          label = "Time To Event Table",
-#'          dataname = 'ADTTE',
-#'          arm_var = choices_selected(c("ARM", "ARMCD"), "ARM"),
-#'          arm_ref_comp = arm_ref_comp,
-#'          paramcd = choices_selected(unique(ADTTE$PARAMCD), "OS"),
-#'          strata_var = choices_selected(c("SEX", "MLIVER"), "SEX"),
-#'          time_points = choices_selected(c(6, 8), 6),
-#'          time_unit = "months",
-#'          event_desc_var = "EVNTDESC"
 #'         )
 #'     )
 #' )
@@ -131,6 +95,7 @@ tm_t_tte <- function(label,
                      arm_ref_comp = NULL,
                      paramcd,
                      strata_var,
+                     conf_int = choices_selected(c(0.8, 0.85, 0.90, 0.95, 0.99, 0.995), 0.95, keep_order = TRUE),
                      time_points,
                      time_unit = "months",
                      event_desc_var = NULL,
@@ -143,6 +108,7 @@ tm_t_tte <- function(label,
   stopifnot(is.choices_selected(paramcd))
   stopifnot(is.choices_selected(strata_var))
   stopifnot(is.choices_selected(time_points))
+  stopifnot(is.choices_selected(conf_int))
 
   args <- as.list(environment())
 
@@ -178,12 +144,16 @@ ui_t_tte <- function(id, ...) {
                           "Select Endpoint",
                           a$paramcd$choices,
                           a$paramcd$selected,
-                          multiple = FALSE),
+                          multiple = FALSE,
+                          fixed = a$paramcd$fixed
+      ),
       optionalSelectInput(ns("arm_var"),
                           "Arm Variable",
                           a$arm_var$choices,
                           a$arm_var$selected,
-                          multiple = FALSE),
+                          multiple = FALSE,
+                          fixed = a$arm_var$fixed
+      ),
       selectInput(ns("ref_arm"),
                   "Reference Group",
                   choices = NULL,
@@ -203,12 +173,29 @@ ui_t_tte <- function(id, ...) {
                           a$strata_var$choices,
                           a$strata_var$selected,
                           multiple = TRUE,
-                          label_help = helpText("from ", tags$code("ADSL"))),
+                          label_help = helpText("from ", tags$code("ADSL")),
+                          fixed = a$strata_var$fixed
+      ),
+      radioButtons(
+        ns("pval_method"),
+        "p-value method",
+        choices = c("wald", "log-rank", "likelihood"),
+        selected = "log-rank"
+      ),
+      optionalSelectInput(ns("conf_int"),
+                          "Level of Confidence",
+                          a$conf_int$choices,
+                          a$conf_int$selected,
+                          multiple = FALSE,
+                          fixed = a$conf_int$fixed
+      ),
       optionalSelectInput(ns("time_points"),
                           "Time Points",
                           a$time_points$choices,
                           a$time_points$selected,
-                          multiple = TRUE),
+                          multiple = TRUE,
+                          fixed = a$time_points$fixed
+      ),
       if (!is.null(a$event_desc_var)) {
         helpText("Event Description Variable: ", tags$code(a$event_desc_var))
       }
@@ -251,13 +238,15 @@ srv_t_tte <- function(input,
     ANL_FILTERED <- datasets$get_data(dataname, reactive = TRUE, filtered = TRUE)
     # nolint end
 
-    paramcd <- input$paramcd
+    paramcd <- input$paramcd # nolint
     strata_var <- input$strata_var
     arm_var <- input$arm_var
     ref_arm <- input$ref_arm
     comp_arm <- input$comp_arm
     combine_comp_arms <- input$combine_comp_arms
     time_points <- as.numeric(input$time_points)
+    pval_method <- input$pval_method # nolint
+    conf_int <- as.numeric(input$conf_int) # nolint
 
     if (length(strata_var) == 0) {
       strata_var <- NULL
@@ -289,8 +278,8 @@ srv_t_tte <- function(input,
 
     chunks_reset(envir = environment())
 
-    adsl_vars <- unique(c("USUBJID", "STUDYID", arm_var, strata_var))
-    anl_vars <- unique(c("USUBJID", "STUDYID", "AVAL", "CNSR", event_desc_var))
+    adsl_vars <- unique(c("USUBJID", "STUDYID", arm_var, strata_var)) # nolint
+    anl_vars <- unique(c("USUBJID", "STUDYID", "AVAL", "CNSR", event_desc_var)) # nolint
 
     ## Now comes the analysis code
     chunks_push(bquote(ref_arm <- .(ref_arm)))
@@ -334,9 +323,13 @@ srv_t_tte <- function(input,
           )
         )),
         data = anl,
+        col_N = table(anl[["ARM"]]),
         event_descr = if (is.null(.(event_desc_var))) NULL else as.factor(anl[[.(event_desc_var)]]),
         time_points = .(time_points),
-        time_unit = .(time_unit)
+        time_unit = .(time_unit),
+        conf_level = .(conf_int),
+        pval_method_coxph = .(pval_method),
+        ties_coxph = "exact"
       )
       tbl
     }))