+++ tests/testthat.R
@@ -1,4 +1,5 @@
 library(testthat)
 library(random.cdisc.data)
 
-test_check("teal.modules.clinical")
+test_results <- test_check("teal.modules.clinical")
+saveRDS(test_results, "unit_testing_results.rds")