+++ tests/testthat/test_nest.R
@@ -2,7 +2,7 @@ library(test.nest)
 library(random.cdisc.data)
 
 
-#test_lintr()
+test_lintr()
 test_strict()
 test_regexp()
 test_spell()